--- base/pre_processing/mpeg_psychoacoustic_masking.py
@@ -1,9 +1,0 @@
-"""
-Compatibility shim.
-
-Psychoacoustic (MPEG-style) masking utilities have moved to
-`base.core_algorithm.psychoacoustic`.
-"""
-
-from base.core_algorithm.psychoacoustic.mpeg_psychoacoustic_masking import *  # noqa: F401,F403
-
