--- base/pre_processing/perceptual_chirp_signal_hd.py
@@ -1,11 +1,0 @@
-"""
-Compatibility shim.
-
-Harmonic distortion / PRB analysis algorithms have moved to
-`base.core_algorithm.harmonic_distortion`.
-"""
-
-from base.core_algorithm.harmonic_distortion.perceptual_chirp_signal_hd import PerceptualChirpSignalHD
-
-__all__ = ["PerceptualChirpSignalHD"]
-
