--- base/pre_processing/psychoacoustic_utils.py
@@ -1,8 +1,0 @@
-"""
-Compatibility shim.
-
-Psychoacoustic utilities have moved to `base.core_algorithm.psychoacoustic`.
-"""
-
-from base.core_algorithm.psychoacoustic.psychoacoustic_utils import *  # noqa: F401,F403
-
