--- base/pre_processing/step_signal_hd.py
@@ -1,11 +1,0 @@
-"""
-Compatibility shim.
-
-Harmonic distortion / PRB analysis algorithms have moved to
-`base.core_algorithm.harmonic_distortion`.
-"""
-
-from base.core_algorithm.harmonic_distortion.step_signal_hd import StepSignalHD
-
-__all__ = ["StepSignalHD"]
-
