+++ scripts/benchmark_prb.py
@@ -17,7 +17,7 @@ os.environ.setdefault("MPLCONFIGDIR", os.path.join(_PROJECT_ROOT, ".mplconfig"))
 if _PROJECT_ROOT not in sys.path:
     sys.path.insert(0, _PROJECT_ROOT)
 
-from base.pre_processing.perceptual_chirp_signal_hd import PerceptualChirpSignalHD
+from base.core_algorithm.harmonic_distortion.perceptual_chirp_signal_hd import PerceptualChirpSignalHD
 
 
 def _parse_harmonic_orders(spec: str) -> List[int]:
